refactored post card styles
